// frontend/app/generated/dtype.ts
export interface components {
    schemas: {
        /** SliderSerializer */
        SliderSerializer: {
            /** @description Non-empty string */
            description: string;
            /** @description Non-empty string */
            image: string;
        };
        /** NewsSerializer */
        NewsSerializer: {
            id: number;
            /** @description String with max 255 chars */
            title: string;
            /** @description Non-empty string */
            description: string;
            /** @description Non-empty string */
            full_description: string;
            /** @description Non-empty string */
            image: string;
            /** Format: date-time */
            created_at: string;
        };
        /** StaffSerializer */
        StaffSerializer: {
            id: number;
            name: string;
            department: components["schemas"]["DepartmentSerializer"];
            role: string;
            image: string;
        };
        /** DepartmentSerializer */
        DepartmentSerializer: {
            id: number;
            name: string;
        };
        /** StudentDataRequest */
        StudentDataRequest: {
            code: string;
            national_id: string;
        };
        /** CollageEmailSerializer */
        CollageEmailSerializer: {
            email: string;
            password: string;
        };
        /** StudentSubjectSerializer */
        StudentSubjectSerializer: {
            subject_name: string;
            subject_code: string;
            sessions: components["schemas"]["SessionInfo"][];
        };
        /** SessionInfo */
        SessionInfo: {
            session_type: string;
            day: string;
            time_start: string;
            time_end: string;
            location: string;
            doctor_name: string | null;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}